+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ### Version 0.6.6
 - [chore] update packages & cleanup & work on improving documentation
 - [js-node/expressjs] minor breaking change. auth callback delimiter from '-' to ';' because '-' can be found in jwt token (https://stackoverflow.com/questions/55201011/what-characters-are-allowed-in-a-jwt-token) and delimiting using '-' can be wrong, use ';' as it can be used in url hash (https://stackoverflow.com/questions/40636281/what-are-the-eligible-characters-in-a-urls-fragment-location-hash) and is not in jwt token
+- [js-node/wip/k8s] moved to [docs]
 - In progress
-  - [chore] update webauthnsimple to v3
+  - [chore] improved fido2 sample in [js-node/expressjs], remove [js-node/wip/fido2]?
   - [js-web/vue-vite] cypress e2e testing
 
 ### Version 0.6.5